changes from lyx -- adding new intro section
It looks like there are a ton of changes but it's actually just adding a new section, and then changing some things related to spacing and reference formatting.
Also adding the bib file, and changing the link to that file.
The changes in the previous commit are actually more substantial than this.